Feat:추가 Feedback

+++ qfit/src/main/java/com/qfit/mvc/model/service/ReviewServiceImpl.java
@@ -1,10 +1,14 @@
 package com.qfit.mvc.model.service;
 
+import java.util.HashMap;
+import java.util.Map;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.qfit.mvc.model.dao.ReviewDao;
 import com.qfit.mvc.model.dto.Review;
+import com.qfit.mvc.model.dto.Review.Difficulty;
 
 @Service
 public class ReviewServiceImpl implements ReviewService {
@@ -22,29 +26,38 @@ public class ReviewServiceImpl implements ReviewService {
 	}
 
 	// 리뷰 작성
-	@Transactional // 트랜잭션 처리
+	@Transactional
 	@Override
 	public void writeReview(Review review) {
 		if (review==null || review.getQuestId()==0 || review.getDifficulty()==null)
 			throw new IllegalArgumentException("Invalid review data.");
+		if (reviewDao.reviewExists(review.getQuestId()) >= 1)
+			throw new IllegalStateException("A review for this quest has already been submitted.");
 		reviewDao.insertReview(review);
 	}
 
 	// 리뷰 삭제
 	@Transactional
 	@Override
-	public boolean removeReview(int questId) {
-		int result = reviewDao.deleteReview(questId);
-		return result==1;
+	public void removeReview(int questId) {
+		if (reviewDao.reviewExists(questId) == 0)
+			throw new IllegalStateException("Review doesn't exist");
+		reviewDao.deleteReview(questId);
 	}
 
 	// 리뷰 수정
 	@Transactional
 	@Override
-	public void modifyReview(Review review) {
-		if (review==null || review.getReviewId()==0)
+	public void modifyReview(int questId, Difficulty difficulty) {
+		if (difficulty==null || questId==0)
 			throw new IllegalArgumentException("Invalid review data.");
-		reviewDao.updateReview(review);
+		if (reviewDao.reviewExists(questId) == 0)
+			throw new IllegalStateException("Review doesn't exist");
+		Map<String, Object> info = new HashMap<String, Object>();
+		info.put("questId", questId);
+		info.put("difficulty", difficulty);
+		
+		reviewDao.updateReview(info);
 	}
 
 }
